update buffer structure and compress data format

+++ packages/contracts/contracts/interfaces/INumbers.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.9;
 
 interface INumbers {
-  function getNum(uint16 num) external view returns (string memory);
+  function getNum(uint256 num) external view returns (string memory);
 }
